🚚 Move secret functions to a subdirectory
