+++ src/functions/secret/index.ts
@@ -1,0 +1,1 @@
+export { SecretFetchForGoogleSecretManager } from './fetch-secret-manager.js';
